+++ src/main/java/com/example/AnketYonetimi/dataAccess/SurveyQuestionRepository.java
@@ -3,5 +3,5 @@ package com.example.AnketYonetimi.dataAccess;
 import com.example.AnketYonetimi.entities.SurveyQuestion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SurveyQuestionRepository extends JpaRepository<SurveyQuestion, Integer> {
+public interface SurveyQuestionRepository extends JpaRepository<SurveyQuestion, Long> {
 }
