+++ src/main/java/com/example/AnketYonetimi/dataAccess/SurveyRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SurveyRepository extends JpaRepository<Survey, Integer> {
-    Optional<Survey> findById(int  surveyId);
+public interface SurveyRepository extends JpaRepository<Survey, Long> {
+    Optional<Survey> findById(Long  surveyId);
 }
